Fixed Socket AsyncRead
Fixed the bug where recv was called once for each raw socket with all
calls after the first returning 0 bytes. Callbacks fire properly now as
a result. When a socket is closed, it's callbacks are also
de-registered. And when the callback list is empty, the worker thread is
shutdown and de-initialized.